[140426] Created class ModuleLoader(require in js), put class MFile in independent files
