--- Modules/Notification/libsubway_notification.so.1
@@ -1,1 +1,0 @@
-libsubway_notification.so.1.0.0
