--- tests/testEvent.html
@@ -1,26 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<HTML>
-<HEAD>
-	<META HTTP-EQUIV="CONTENT-TYPE" CONTENT="text/html; charset=utf-8">
-	<TITLE></TITLE>
-	<meta name="generator" content="Bluefish 2.2.4" >
-	<META NAME="CREATED" CONTENT="20130719;1421100">
-	<META NAME="CHANGED" CONTENT="0;0">
-	<STYLE TYPE="text/css">
-	</STYLE>
-	<script type="text/javascript" >
-		function onResizeEvent(a,b,c,d)
-		{
-			UNIX.SendNotify("OnResize:" + a + "," + b + "," + c + "," + d,"qt4logo");
-		}
-		function onKeyPressEvent(a)
-		{
-			UNIX.SendNotify("OnKeyPress:" + a,"qt4logo");
-		}
-	</script>
-</HEAD>
-<BODY LANG="zh-CN" DIR="LTR">
-<P STYLE="margin-bottom: 0cm"><BR>
-</P>
-</BODY>
-</HTML>
