+++ tests/testMFile-obsolute.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html>
+<head>
+  <meta name="subway_title" content="MFile test" />
+  <meta name="subway_size" content="320x240" />
+  <script type="text/javascript">
+  function init(){
+	MFile = require.LoadModule("MFile");
+    MFile.DownLoad("https://github.com/EasternHeart/Cubway","cubway-index.html");
+  }
+
+  </script>
+</head>
+<body onload="init()">
+</body>
+</html>
